configure role access, fix some bugs in services

+++ BusinessService/src/main/java/com/company/innowise/airticketsapp/businessservice/controller/TicketController.java
@@ -31,7 +31,7 @@ public class TicketController {
     }
 
     @GetMapping("/{id}")
-    public TicketDto getTicket(@PathVariable Integer id) {
+    public TicketDto getTicket(@PathVariable Long id) {
         return ticketService.getTicket(id);
     }
